+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Minor fixes to autocrafting (raoulvdberge)
 - Fixed Exporter in Regulator Mode not regulating properly when same item is specified multiple times (raoulvdberge)
 - Fixed air appearing in Grid (raoulvdberge)
+- Fixed config categories not correctly appearing in ingame config GUI (raoulvdberge)
 
 ### 1.5.25
 - Fixed not being able to autocraft different Storage Drawers' wood drawers (raoulvdberge)